🐥 Support `decode` command, decode `*.wasm` to wat

+++ projects/legion-app/src/cli_cmds/cmd_encode.rs
@@ -8,17 +8,22 @@ pub struct EncodeCommand {
     /// output wasm file name
     #[arg(short, long)]
     output: Option<String>,
+    /// dry run
+    #[arg(long)]
+    dry_run: bool,
 }
 
 impl EncodeCommand {
-    pub async fn run(&self) -> anyhow::Result<()> {
+    pub async fn run(&self, args: &LegionOptions) -> anyhow::Result<()> {
         let input = Path::new(&self.input);
         let mut parser = wat::Parser::new();
         parser.generate_dwarf(GenerateDwarf::Full);
         let wasm_bytes = parser.parse_file(input)?;
         let output = self.make_output_name(input).await?;
-        let mut wasm = File::create(output).await?;
-        wasm.write_all(&wasm_bytes).await?;
+        if !self.dry_run {
+            let mut wasm = File::create(output).await?;
+            wasm.write_all(&wasm_bytes).await?;
+        }
         Ok(())
     }
